Standardise EHSH_LOG_* format strings.
Before, we had a random mix of newline and no newline at the end of the format
strings passed into the EHSH_LOG_ macros. I've standardised on *not* having the
final newline. If some logger requires this, it can always add it itself. The
net result is that log output on stdio doesn't have lots of random orphaned close
double bracket lines after (almost) every log output line.

Note: various files touched by the ongoing Qt integration are *not* fixed in this
commit, as they are going to be addressed in a following commit.