Lots more work on Qt integration
Note: this version includes a nasty hack to 'create' the widget instance for
our text box. I cannot figure out why the application itself never causes the
create function to be called.

Due to the hacky plumbing, when we click on the button in the GUI, there is
currently a segfault in the widget's event_handler callback. Further, deeper,
investigation is required to understand how the widget creation is supposed to
happen in a "MODE_B" graphics build in order to ensure all of the event
plumbing is correctly in place.